Magic key placement.
This should be the final piece of the puzzle for magic chest
implementation. To be able to unlock a magic chest (they spawn locked by
default), you need a magic key. Magic keys will not spawn at random,
which means they also cannot be had via polypiling. They are made of
gemstone, so are inherently fooproof. Only three at most will ever
exist in the game (if Fort Ludios does not spawn in a given game, the
number drops to two). The locations of these magic keys are as follows:

* Gollum's Cave
* Mines' End
* Fort Ludios

Gollum's Cave - the location of the magic key is inside the chest in the
center island. While placing the magic key here, I did a bit of a revamp
of that level, amking the shallow water pathways a bit random. There's
now also a small chance of shallow water placement from the center
island to the far island that would allow access via jumping.

Mines' End - finding the magic key here will be the most fun in my
opinion. In four of the five Mines' End levels, the magic key is buried
inside of a treasure chest in a semi-random spot on that level. An
engraving will also be on that spot to give the player a hint. In the
'Sewers of Waterdeep' level with the Rat King, the magic key is located
within the regular chest which resides in the treasure room. In the
'Gnome Kings Wine Cellar' level, the magic key is buried, but one of the
random locales is another small cavern surrounded by rock, much like
where the luckstone and gems are hidden, and can be easily reached by
digging.

Fort Ludios - on this level, the magic key is locked away within an iron
safe which is placed within the zoo near the entrance portal.

Players will need to adventure a bit and find one of the keys before
they can use the magic chests 😃