Change placement of magic key in Gollum's Cave.
After playing a few games, it became apparent that having the magic key
in the center island chest made it too easy to obtain. It's now located
in the chest on the far island. I did increase the odds of shallow water
forming allowing a jump path to the far island.