Fix: zap/explosion types, overhaul.
With the inclusion of a new breath attack (shimmering dragon/stun), the
way the game determined the correct effect type for breath
attacks/spells/wands/explosions had to be readjusted. It became a bit
confusing, as some 'magic numbers' were hard-coded, and if you changed
one to make that effect type correct, it could then break another effect
elsewhere. qt stepped in and refactored how the game handles these types
(look for a PR for vanilla NetHack soon I hope), so in the future if new
types are added, it should make it much easier to incorporate.

Co-authored-by: entrez <me@entrez.cc>