Fix: Give message when walking onto a magic chest square.
Loggers_VIII pointed out it felt weird step on a magic chest square and get no message. I agree. Give a message akin to the one for objects. Its feedback is a bit odd if you walk on the square while blind if it has objects on it. That seems rarer than an empty magic chest square. Besides, all these coding sins will be washed away in the second coming of magic-chests-as-objects, hopefully.