Fix: offhand warning
Loggers_VIII discovered that 'a'pplying a pick-axe (or any other tool that you wield) while wielding a weapon that gives warning and having another weapon in the offhand slot that gives warning would cause the offhand weapon to start warning you of things even though it wasn't wielded. This was due to an oversight in set_artifact_intrinsic---it turned on warning for any item with owornmask, even W_SWAPWEP.

In the course of testing, I noticed that a seemingly similar issue happened if you were twoweaponing when you applied the pick-axe, but it was actually a separate bug caused by not turning off offhand intrinsics when stopping twoweaponing.