Fix: Allow non-chaotics to wield the Sword of Kas again
Now that artifact blasting is back to vanilla behavior, the Sword of Kas cannot be handled by cross-aligned characters, which it is considered a bane to. This is, in fact, also vanilla behavior, but (I think) not intended in the case of the Sword of Kas, so make it a special case.

Also, change death messages a little so that the material of a weapon that blasts you to death is only mentioned if relevant. (It is also not mentioned if the weapon happened also to be a bane, since that seems like the more salient characteristic).