Fix: Display suffocation attacks differently from drowning
Loggers was surprised when a shambling horror with an "engulf drown" attack worked on a tortle. Turns out gelatinous cubes also display as "engulf drown," which isn't great.

Speaking of "isn't great," my solution was just to add a special case in the moster lookup code. Shambling horrors suffocate, because 1) that's how the current mhitu code works, and 2) they probably aren't made of water.