Wand of Orcus changes.
Previously, the Wand of Orcus was deemed too powerful for any player to
be able to wield. This commit does away with that restriction. Players
(and non-demonic monsters as well) can now wield the Wand of Orcus, but
being in control of such deadly magic comes with a price. The changes:

* when wielding the Wand, health regeneration is completely suppressed
(same as being in the Valley of the Dead)
* wielding the Wand, if not already cursed, will become cursed and weld
itself to the players' hand (if not an Infidel)
* if the Wand is welded and is uncursed/blessed, there is a random
chance it will re-curse itself
* dropping the Wand from inventory, if not already cursed, will curse it
(same behavior as dropping a loadstone)
* maximum hit point drain vs monsters is now more inline with how it
affects the player
* the Wand can be invoked to tame any nearby undead

Not pushing this to the servers just yet, I'd like some review on this,
see if anything else needs to be added or adjusted. I'm thinking the
Wand should not be able to be dual-wielded with another artifact. Max
hit point damage vs monster could probably be tweaked too.

Also in this commit - had to remove the status_hilite entries from the
sysconf tweak, as those options apparently override player configs.