Fix: Wand of Orcus dual-wielding and invoke function.
The Wand of Orcus' invoke power was meant to tame, or at the very least
make peaceful if possible, any nearby undead. Using maybe_tame() took
the BUC status of the artifact into account, and as its natural state is
to be cursed, would wind up making any undead angry. So switch
maybe_tame() to tamedog().

Fix the twoweaponing characteristics when the Wand is factored in. Final
behavior is thus:

* the Wand can never be twoweaponed with another artifact
* the Wand can be twoweaponed with another regular weapon, but only if
the Wand is the primary weapon; it will not tolerate being held second