Fix: various artifact death messages.
A prior commit negated the feedback for when the Wand of Orcus outright
killed its target with death magic. Instead of just putting things back
to normal, I took this opportunity to provide specific relevant feedback
based on the event. This started with the Wand, but bled over into
invoking the Eye/Hand of Vecna, and Vecna's death gaze.