Don't allow supermaterials to be erodeproofed
This led to oddities like forged weapons not stacking despite displaying identically in inventory and incentives to erodeproof dragonhide armors. Instead, make it so that supermaterials never have the oerodeproof bit set to TRUE.

Another solution would be to display fixed status for supermaterials and allow them to be fixed, but the only time oerodeproof is relevant for supermaterials is for the destroy armor monster spell, so that seems like an overreaction.

This commit makes it so that supermaterials can only ever take three shots from the destroy armor spell, while fixed erodeable materials can take four. This is a little odd, but it seems reasonable. And a downside to dragonhide isn't so bad, I think.