Tweaks to dlords following levelports
The "ominous presence" message when leaving a level with a demon lord is removed according to player feedback; it was more confusing than helpful. I decided to make the arrival message a bit more obtuse as well, in keeping with the Nethack spirit.

More substantively, demon lords will join the migration list if left behind on *any* Gehennom level; I realized it was silly to only allow it on lair levels.