Pacify Kathryn the Enchantress with a candy bar.
Mini-me told me a long time ago, if she's ever feeling sad or upset,
just give her a candy bar and she'll be her usual happy self again.
While this doesn't always work for every situation, it has worked in the
past. I've mentioned this to the people that hang out in #evilhack on
IRC, and they all thought that would wonderful to see that make its way
into the game.

So. Here we are. Once Kathryn the Ice Queen has been defeated, the curse
removed, and she transforms into the peaceful Kathryn the Enchantress -
if you then make her angry, she can be pacified by tossing her a candy
bar. This will work for every role except for Infidels, as in the
presence of them she never becomes peaceful to begin with, and never
will be.

Co-authored-by: entrez <me@entrez.cc>