Add option to alter behavior of 'A' when looting
Many players (like myself) aren't fans of the autoselect all option when looting, since it makes it easy to dump explodey stuff into a BoH. This commit adds an option to slightly alter the behavior of the 'A' key---if enabled, selecting 'A' will take you to a menu with everything selected by default.

This option was inspired by what I thought 3.7 did, but I was entirely wrong. I wanna say Noisytoot came up with the idea? Or maybe described how some variant handles it?

If intentionally selecting 'A', then the player's actions are very similar (just a quick double-press of the enter key instead of one after hitting 'A'). But, if the selection was accidental (e.g., was holding down shift in preparation for hitting 'X', then changed your mind and went to hit 'a', or meant to hit 'X,<enter>' but missed the first keystroke and hit ',<enter>' instead, you will have a chance to abort before dumping your stuff in the bag.