Fix: impossible() when killing multiple vault guards in succession
Killing two more guards in a row (while they were in their corridor) would result in place_monster complaining about placing guard on top of itself at (0,0). It seemed to be that the new guard would be stored at the same address as the previous one, and the old pointer was never removed from level.monsters[0][0]. Not sure why the same address kept being used, but this seems to fix it regardless.