Fix: killing quest leader won't flag the quest as complete if they were killed while polymorphed.
Was recently discovered that if you kill your quest leader while they
are polymorphed into another form, the act of killing them will not flag
the quest as completed. This bug almost cost tau their game, but luckily
they were able to exploit a janky workaround and finish it out.

To fix - change the critera of what the game looks for when the quest
leader is killed. There's a few other places where this change should be
used, but for now lets just fix the potential game-ending bug. Thanks
terrain for seeing this solution quickly.