Fix: Checking floor for tippables can stone you
Feeling around for containers will stone you if gloveless and blind, like the behavior of #loot. 'm#tip' will skip floor objects, again like with #loot. This technically prevents someone from playing with menustyle:traditional and then requesting a full menu, but why would anyone ever do that? Of greater concern is the opacity of using menu_requested like this, but if it's good enough for vanilla's loot functions it's good enough for me.