Fix: prevent the Amulet of Yendor from being teleported whilst in the Sanctum.
If the high priest in the Sanctum is killed, and they then drop the
Amulet of Yendor, and then if it is teleported away (spell/wand of
teleportation) before the player has picked it up, there's a chance it
could teleport to the west side of the map. If that happens, it can't be
retrieved, because it has to be picked up for the gates to form to allow
access to that side of the map. This can create an unwinnable situation.

This hasn't actually happened during an online game, but during a recent
game today, we thought it *might* have happened. Tested out the
hypothesis and sure enough, it was possible.