Fix: confused enchanting dragon scales
The dtsund-DSM patch made it so that confused enchant armor would polymorph you into a dragon if wearing dragon scales. Between adding the ability to choose which armor piece to enchant, refactoring that addition, and adding code to prevent fooproofing supermaterials, this particular behavior was overridden by a general confused-enchant-armor case. Restore the original dtsund behavior (if the dragon scales are chosen/the scroll is not IDed).