Fix: only cause light damage when sources begin burning near you
This may be the cause of mobileuser's "Ow, the light hurts!" issues, or at least a contributing factor. Specifically, I think it would damage drow characters if moving to a new level where a monster with gold scales/etc. was generated (which is somewhat likely, especially on Astral). However, it's doubtful that several such monsters would spawn, so there may be another cause of the plethora of messages.

Make it so new light sources only damage drow characters if the light reaches them (or at least, if they have line-of-sight to the light. I don't think LOS is symmetrical?). Also, make it damage them even when wielding the Staff of the Archmagi. While it would be reasonable to make SotA prevent light damage, it only did so against artifact_light sources, not magic lamps. So, I assume it was unintended.