Require manual suppression of attacks vs. dangerous enemies
Auxiliary attacks when not polymorphed (tentacles for illithids, stings for demons, and kicks for monks) are frustrating when fighting something with a dangerous passive attack, like a footrice or black dragon. So, they were being automatically suppressed when fighting such enemies. However, the method was ad hoc and led to strange situations:
-The monk situation was odd because it cut across races, some of who could wear shoes (and did not get kick suppression), and some that couldn't (and did). If you were a tortle monk, for instance, all you needed to fight a black dragon was Dragonbane and a can-do attitude. However, human (for example) monks did not have their kick attack suppressed, since they could in theory be wearing boots. In practice, you would lose the boots to disintegration and then be screwed.
-The other auxiliary attack cases were also confusing. Both illithid and demon player races have closely analogous polyforms, for who tentacle/sting attacks were similarly suppressed. But, other polyforms were not so lucky. Foocubi's bites were never suppressed (though vampires' bites were?). Centaur kicks were unhindered, which was especially confusing given the existence of the player race (which does not have a kick attack).
-Also, a bug with fall-throughs in the attack type switch statement meant that, depending on the specific attacks of a polyform, some other attacks might be skipped as well. For instance, vampires had both claw and bite attacks suppressed.

Furthermore, the whole "auto-suppression" idea seems a bit outside the Evilhack ethos---paranoid water's existence is a serious and ongoing debate, yet some characters could blithely waltz around cockatrices with no issues.

With this commit:
-all automatic suppression of attacks is removed
-only four types of attacks can be suppressed:
  -an illithid player's tentacle attacks (but not a mind flayer polyform's)
  -a demon player's sting attack (but not e.g. a horned demon's)
  -a monk's martial arts kick attack, regardless of player race
    -note that if polymorphed into something with an AT_KICK attack, those kicks still cannot be suppressed
  -shield bashes
-when applicable, attacks can be suppressed by forcefighting with 'F'

Hopefully, the narrower range of which attacks can be suppressed will make things easier to remember. It also restores a major downside to some polyforms, which I think makes the choice to, for instance, go for a permanent master mind flayer or foocubus polyform more interesting, as well has hewing more closely to vanilla.

As a helpful side effect, this prevents tentacle attacks and shield bashes when stealing as a rogue or convict.