Fix: fixedness knowledge preventing merging of supermaterials
The fixedness of supermaterials now how no impact on gameplay, and they are always unfixed. However, the *knowledge* of whether a supermaterial item was fixed or not still mattered for merging purposes. To exacerbate things, whether or not this status was known did not matter for the purposes of considering it fully identified or not.

bouquet noticed this with two stacks of apparently identical, fully-identified mithril daggers that would not merge.

To repeat the bug: wish for two mithril daggers. Drop them on an altar. Read a blessed scroll of magic detection. Wield one of the daggers and kill grid bugs until the enchantment is known. Read a scroll of identify. The other dagger will be fully identified (the one you're wielding can't be IDed further), and they will not merge.

This fixes that, and I also tweaked a few lines of code I came across in the process to make their meanings clearer.