Restore auto-suppression of some dangerous attacks
This partially reverts commit cad8389 and restores automatic suppression of most deadly auxiliary attacks. Rejoice! The attacks that can be suppressed with force-fight (and only those attacks) will be automatically suppressed if particularly deleterious effects would happen and there is no way to prevent it by wearing armor. This essentially means that, so long as you are not polymorphed, things are as they were before these recent commits. Force-fight manual suppression has not been changed from the previous commit. The removal of suppression for all polyforms remains in effect from last commit.

For clarity, automatic suppression happens if you are not polymorphed and:
- you're an illithid making a tentacle attack against something which can passively stone, slime, or fully disintegrate you, or a Rider, or a creature of the same species
- you're a tortle or centaur monk making a kick attack against something which can stone or fully disintegrate you
  - note that you can still suppress the attack manually if e.g. you're a human and don't want disintegrated boots
- you're a demonic infidel making a sting attack against something which can stone or fully disintegrate you

On the code side, kick suppression was moved to uhitm.c. Now you can kick a black dragon as a non--disintegration-resistant tortle/centaur monk if you want to play Evilhack-style Russian roulette. Rejoice? Force-fight manual suppression remains, both because it is useful for monks and satiated illithids, and because it fixes the thievery bugs mentioned last commit and fixing them in a more targeted manner looks kinda hard.