Fix: crash when examining discoveries with '`' in certain cases
The curses version of yn_function could return whitespace characters if the default was '\0'. This leads to a crash if you use the '`' option with curses enabled and a traditional/combination menu style (see doclassdisco() in o_init.c). Allow '\0' to be returned as a default. The tty version allows this, so it should not lead to unexpected input for other yn_function calls (I tested several to be sure).