Fix: Amulets could burn your skin but then allow you to regenerate
Noticed that a mithril amulet caused material hatred damage to orcs when worn but would not prevent regeneration if wearing a shirt or armor. will_touch_object wasn't quite consistent with the check in allmain.c for whether amulets touched skin or not. Update it to match allmain's version: a shirt or body armor will prevent an amulet from burning you if it is made of a hated material.