Fix: unique monsters are not bound by maximum hit point value in grow_up().
Unique monsters are not bound by a hard limit to maximum hit points that
can be gained, just as the player isn't. It was noticed in the recent
game that Lucifer, who starts with 666 hit points, drank a potion of
gain level and had his maximum/current hit points set to 400. There are
a few unique monsters in EvilHack that start with more than 400 hit
points, thus this hard limit should not be imposed.