Fix: artifact rings have a fixed material.
Because rings inherently have a random appearance/material, many times
the two artifact rings in the game (the One Ring [gold] and the Ring of
P'hul [gemstone]) would spawn exhibiting the material properties of its
base item, but would still appear as gold or gemstone, etc. Address the
confusion, and make sure the base materials for each artifact ring are
made of the material they're supposed to be.

This does set up a condition to where an observant player could quickly
learn that if, for example, they see a gold ring, they'll know it's
invisibility without having to formally ID it. To get around that is a
bit more work than I'm willing to do at the moment.