Fix: names of new ring types when type is known
When ancient/lustrous rings were IDed, they'd show as "ring of ancient/lustrous". Display them as "ancient/lustrous ring" instead.

This doesn't fix the issue that object lookup only triggers correctly if you search "ancient" or "ring of ancient." The rings' OBJ_NAM's are "ancient"/"lustrous" respectively, and the object lookup uses name_to_otyp() which only chops off the name of the class of the object if it comes first. Similarly, you can't look up "invisibility ring," even though it's wishable. But in that case the official name is "ring of invisibility" so you wouldn't normally try to look that up anyway.