Fix: monster priests shouldn't wield edged weapons either
Monster priests would wield an edged weapon if it was better than their starting mace. Forbid them from doing so.

The Archbishop of Moloch is exempt because he's an infidel. The high priest of Moloch, on the other hand, is old-school and wants no truck with newfangled infidel traditions.

Also, I cleaned up some redundant Grunthack code. It was hurting anything, but only because it was literally useless.