'Deslog' the first wizard's tower level.
The final big room before reaching the upstairs in the Wizard's tower
(first level) was one giant zoo. Kind of turned that space into a
time-sucking slog fest. What this commit does, is get rid of the zoo,
throw in a some lava, and add a number of fire/lava-loving monsters in
that space instead. Still a challenge, but quite a bit fewer monsters to
have to wade through.

Also adjusted the placement of the one guaranteed player monster,
appears he was stuck in a tree.