New trap: ice.
This commit adds a new type of trap - an ice trap. The coding here is
somewhat of a mix between xNetHack's and Hack'EM's implementation, with
further modifications after the fact.

An ice trap behaves mostly as a fire trap does - it's stepped on, and a
blast of freezing mist shoots upwards from the trap, damaging whatever
stepped on it, and potentially freezing any potions in open inventory.
The amount of damage can be significant (4d8, can be reduced depending
on cold resistance percentage), but does not reduce max hit points like
a fire trap can sometimes do.

Where this trap really stands out: if the player stepping on it has
cold resistance, but not inately, and their resistance is 50% or
greater, there's a chance they can have an amount up to 50% permanently
reduced (for monsters, their cold resistance can be taken away
entirely). If the player has inate cold resistance (gained by role or
race), they are temporarily stricken with cold vulnerability for a
period of turns. Having an extrinsic source of cold resistance is the
only way to maintain partial (read: not inate) resistance if stepping on
an ice trap.

Various levels have had ice traps added, notably the Ice Qeen's realm,
as well as a few choice areas in Vecna's domain, Gehennom, and a couple
others. Ice traps will not spawn randomly outside of the Ice Queen's
realm. Also note, magic traps can behave as an ice trap randomly, just
as they can as a fire trap.