Fix: certain monster steeds should only be ridden by certain monster riders.
For the player, only as a Drow can you ride spiders or cave lizards, or
only as an Orc can you ride a warg, and so on. Monster rider/steed pairs
didn't have this restriction, which seemed incorrect. This commit
addresses this, and brings monster rider/steed pairs to be more inline
with what the player can ride or not ride.