Lolth will be peaceful towards Drow players under the right circumstances.
If the player is Drow, stay true to their alignment (chaotic), and are
at least fervently aligned (9 or greater), Lolth will be peaceful
towards them.

This now makes me think that under the same conditions, if a chaotic
Drow player kills Lolth, there should be some sort of harsh penalty,
seeing as how she is their deity and all. Hmm.