Attempt to make Wizards less reliant on weapons, focus more on spellcasting (round one).
This is my initial attempt to make Wizards want to be more reliant on
their spells instead of being forced to 'tank' and melee most times.

Playing Dungeons & Dragons, Wizards (Magic-Users as they're called in
ad&d) would almost never melee. The combat arms types in the group would
handle that aspect of battle, while they stayed back and cast spells in
support of the rest of the group. As the Wizard increased in experience
and abilities, they got to the point where they didn't need the 'support
staff' nearly as often, if at all. Granted, Wizards in NetHack don't
have a party of warriors and healers to rely on starting out. However,
they wind up fighting melee style quite a bit, even at higher levels.
This commit is an attempt to change that a bit.

The changes:

* Wizards casting the spell 'force bolt' will only cost them one point
of spell power per cast, instead of the usual five points
* Wizards will not start with any weapons at all
* Wizards will start with an extra wand, and that wand will always be a
wand of magic missile

Player monster Wizards are changed in the same way (no weapon, extra
wand of magic missile).

These are very simple changes, but the effect could be profound.
Definitely want to play test thoroughly before the release of this
version.