Changes to to-hit formula for force bolt spell.
Wands of striking and force bolt spell have a fixed chance to-hit its
target, which is separate from spell_hit_bonus(). The default formula
if not engulfed is rnd(20) < 10 + find_mac(mtmp), meaning that if your
target has an AC -9 or better, you'll never land a hit with either wand
or spell. Wasn't ever really that big of a deal in vanilla NetHack,
because you'd never see a monster with an AC better than -10
(disenchanter), and by then you'd have better means of dealing with
something like that. That's not the case in EvilHack - this formula
makes wands of striking/force bolt spell almost useless in the
mid-to-late game, even in regular NetHack to a certain extent.

This commit now factors in your skill in attack spells when using force
bolt spell. At basic skill, the odds are the same as default. At
skilled, the odds go up significantly, and at expert you'll land a hit
every time on anything that has an AC of -10 or worse.

I may tweak this further, but we'll play with this change for now.