New spell - critical healing.
Critical healing spell - a level 6 healing spell that will heal 6d12 of
damage if cast at basic level, and scales up if skilled/expert in
healing magic. There is a 25% chance that the target of this spell will
also be affected the same as if 'cure sickness' spell were cast on them.