Allow Illithids with intrinsic flying to enter into and stay in pools/moats.
Much like the Monk with intrinsic water walking commit (0d4dce6), this
commit allows Illithids that have gained intrinsic flying to go for a
swim, and they'll stay at the bottom of the pool if they have a means of
breathing underwater.

There's a fix here as well. I noticed that when flying or having
intrinsic water walking, you couldn't get back out of the water unless
you climbed out to adjacent dry land. That did not feel right, so now
you can use the up command to step up/fly out of the water.