New race: Draugr (initial commit).
A couple days ago, I was asked if vampire race was ever going to be
added to EvilHack. I said no, because it's already in several variants,
it's buggy, and so on. I jokingly alluded to adding a playable zombie
race since EvilHack is kinda known for the 'zombie apocalypse'. And then
I thought about it a bit. And then some more, and now after a late night
coding session and some more today, here we are.

Draugr - a Norse/Viking term for 'an undead creature from the
Scandinavian saga literature and folktales'. Draugr have a lot in common
with zombies, but are not completely mindless. The highlights of this
race - undead/zombie, can revive if 'killed' x number of times (x being
random each game), has a zombie sickness bite attack, can never learn
or cast spells (almost mindless but not quite) and all of the benefits
and weaknesses associated with being undead.

I'm leaving out a lot of details here for what's in this commit and what
I have plans for in the near future. In its current form, a Draugr will
revive indefinitely, so know that going in if you compile this and play
around with it. I'll add the revive tracker and threshold bits soon. I'm
also not entirely satisfied with the role combos, so more to do there.
Consider this a rough draft.

Any comments or suggestions, feel free to share.