Shop prices for Draugr, other price tweaks.
Draugr are permanently banned from entering a shop via normal means, but
this doesn't mean they couldn't wind up inside a shop anyways. Could
teleport in, or fall through a trap door into one, and so on. The
capitalist shopkeeper mentality is 'well, so long as they're here...'
but the markup for zombie types is extremely high. Twenty times that of
normal for most shopkeepers. While here, I also made the pricing for
gnome and nymph shopkeepers a bit more varied.