Fix: ^X message for being warned of undead
With the previous MH_UNDEAD commit, the ^X screen counted MH_ZOMBIE and MH_ZOMBIE separately, which looked odd. Only say the player is warned of "undead" unless specifically zombies are warned of (not currently a thing for any artifact).

Also, fix up role.c which somehow got skipped in the last merge.