Fix: summoned spheres could zombify victims
Since exploding summoned spheres go through mon_xkilled -> xkilled,
which assessed possible zombification of the killed monster based on the
state of the hero, the explosion of a summoned sphere could end up
creating zombies if its summoner was in a state to do so herself (due to
polymorph or race, not wielding a weapon, etc).  Add a bit to
xkill_flags to indicate xkilled is being used in this scenario, and
assess zombification in that case based on the existing value of the
'zombify' global set in mhitm.c, as happens with normal mon vs mon
control flow that goes through monkilled().