Update xkill_flags comment
Add XKILL_INDIRECT, and there was no need to give the specific values of
the macros anyway compared to their names.  I don't really remember if
this is the proper "K&R era NetHack style" way to format a multiline
comment on a function parameter, but hopefully it makes sense.