Fix: zombie_maker and hero or monster race
zombie_maker wasn't checking mplayer/monster race, and the attempt to
check the hero's race was inappropriately applying the hero's draugr
race to all non-zombie monsters.  This ends up turning it into just a
wrapper for the racial_zombie macro, but I left it as a function in case
there needs to be more complex logic or special exceptions introduced in
there in the future.