Fix: some bugs with various life saving conditions.
If a Draugr is also wearing an amulet of life saving, don't run through
the intrinsic revival routine.

Bigger bug here though, was the bit checking for Draugr and nonliving in
general, it was checking for 'non' nonliving instead. Put a ! where it
wasn't needed. That could have been bad 😁. Cleaned up some feedback
here as well.