Fix: no magic resistance/reflection conducts were being broken by monsters.
When making these conducts, and handling worn/wielded objects, I forgot
that armor_provides_extrinsic() was used by monsters as well as the
player (that function really needs to be renamed, but that's for another
day). To fix, move the conduct checks to setworn(), which is used
exclusively by the player.