Draugr are properly vulnerable to fire.
Many undead monsters have the M4_VULNERABLE_FIRE flag, meaning that they
take 1.5 times the amount of damage from any fire-based attack. That
flag does not carry over to the player. Using the HVulnerable_fire
intrinsic flag would cause the player to essentially start with -50%
fire resistance, and would never go above 50% even with an extrinsic
source of fire resistance. So that's a no-go.

To express their vulnerability to fire attacks properly, the damage
multipler is added to the resist_reduce() routine, and will scale
properly based on the players fire resistance.