Fix: tty character selection filtering menu
Several letters were used twice or more by various options---the old code assumes that no roles start with the same letter and that now races/alignments/genders start with the same letter, and between caveperson/convict and a whole bunch of races this is no longer true. Hardcode the stuff instead. If lots more races/roles are plans, perhaps this stuff should be made a field in the race/role constructs, but for now this suffices.