Fix: blunt weptools had +4 TH bonus
It used to be (like in vanilla 3.3 or something)  WHACK/PIERCE/SLASH were defined as 0/1/2. The damage type for weptools was used as their to-hit bonus. But, when WHACK was redefined as 4, the to-hit bonuses changed, too. Pretty sure this is still extant in vanilla as well, just less noticeable because vanilla has no object lookup.