Fix: consume_tin() bypassed is_edible()
Draugr race cannot eat anything except non-veggie corpses and eggs,
meaning no processed foods, which includes anything in a tin. Players
could bypass that rule by using a tin opener to open a tin. This commit
closes that loophole, and forces consume_tin() to check whether the
contents are edible or not.

This isn't the most elegant solution - currently this treats all tins in
any state as off-limits to Draugr. Which is fine for now, because
any food found within a tin is technically processed/preserved. In the
future, I think it'd be neat to allow Draugr to eat meat from a tin if
its contents were rotten (side note: food sense [from spell/as a Hobbit]
should be able to tell if the food inside a tin is rotten or not before
eating it).