Forcefighting a door or any type of tree while burning hands spell is active can burn them down.
Doors (for the time being) and trees are made of wood. Fire burns wood.
Touching either while the spell 'burning hands' is active should affect
these things accordingly.

Forcefighting a live tree with burning hands will burn it up and turn it
into a dead tree. Doing the same to a dead tree makes it burn away
completely. Exception to this rule is if these trees are 'petrified' (no
dig level), in which case they remain intact.

Forcefighting a door with burning hands will cause the door to burn away
completely. If the door was a door to a shopkeepers shop, the shopkeeper
will become angry, but it can be pacified if you pay for the cost of
repairs. Exception to this rule are the doors in Sokoban that lead to
the prizes found there.